Fix Bug

+++ Application/Features/Files/Commands/CreateUpload/CreateUploadCommand.cs
@@ -2,6 +2,6 @@ using Application.Features.Files.Dtos;
 
 namespace Application.Features.Files.Commands.CreateUpload
 {
-    public sealed record CreateUploadCommand(string Title, long Size, int ThumbnailTime, int Duration)
+    public sealed record CreateUploadCommand(string Title, long Size, int ThumbnailTime, double Duration)
         : IRequest<OneOf<CreateUploadDto, Error>>;
 }
